backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
azure-cosmos>=4.5.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
httpx>=0.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
